Cleanup: Reuse similar code for survey related views.
