+++ docs/changes.rst
@@ -4,6 +4,8 @@ Changelog
 15.1.0 (unreleased)
 -------------------
 
+- Cleanup: Reuse similar code for survey related views.
+  [thet]
 - Fix incorrect named entity in configure.zcml.
   [thet]
 - Use NuPlone 2.2.2.
